pip/myhand/requirements.txt
argparse
requests
openai==1.3.4
prompt_toolkit
Pygments
datetime
netifaces
geocoder
googlesearch-python
art
apsw
gTTS
google-cloud-texttospeech
yt-dlp
pyperclip
colorama
tiktoken
docker
unstructured[all-docs]
pillow
pyautogen[retrievechat,teachable,mathchat]==0.2.0b5
chromadb
pygame
